+++ src/braket/circuits/translations.py
@@ -67,6 +67,7 @@ BRAKET_GATES = {
     "cswap": braket_gates.CSwap,
     "gpi": braket_gates.GPi,
     "gpi2": braket_gates.GPi2,
+    "prx": braket_gates.PRx,
     "ms": braket_gates.MS,
     "unitary": braket_gates.Unitary,
 }