+++ src/braket/parametric/free_parameter.py
@@ -20,6 +20,69 @@ from sympy import Symbol
 
 from braket.parametric.free_parameter_expression import FreeParameterExpression
 
+PREDEFINED_VARIABLE_NAMES = {"b", "q"}
+
+# The reserved words are picked from below
+# https://github.com/openqasm/openqasm/blob/main/source/grammar/qasm3Lexer.g4
+# https://github.com/openqasm/openpulse-python/blob/main/source/grammar/openpulseLexer.g4
+QASM_RESERVED_WORDS = {
+    "OPENQASM",
+    "include",
+    "defcalgrammar",
+    "def",
+    "cal",
+    "defcal",
+    "gate",
+    "extern",
+    "box",
+    "let",
+    "break",
+    "continue",
+    "if",
+    "else",
+    "end",
+    "return",
+    "for",
+    "while",
+    "in",
+    "pragma",
+    "input",
+    "output",
+    "const",
+    "readonly",
+    "mutable",
+    "qreg",
+    "qubit",
+    "creg",
+    "bool",
+    "bit",
+    "int",
+    "uint",
+    "float",
+    "angle",
+    "complex",
+    "array",
+    "void",
+    "duration",
+    "stretch",
+    "gphase",
+    "inv",
+    "pow",
+    "ctrl",
+    "negctrl",
+    "dim",
+    "durationof",
+    "delay",
+    "reset",
+    "measure",
+    "barrier",
+    "true",
+    "false",
+    "waveform",
+    "port",
+    "frame",
+}
+
 
 class FreeParameter(FreeParameterExpression):
     """Class 'FreeParameter'
@@ -66,7 +129,7 @@ class FreeParameter(FreeParameterExpression):
             Union[FreeParameter, Number]: The substituted value if this parameter is in
             parameter_values, otherwise returns self
         """
-        return parameter_values[self.name] if self.name in parameter_values else self
+        return parameter_values.get(self.name, self)
 
     def __str__(self):
         return str(self.name)
@@ -92,8 +155,18 @@ class FreeParameter(FreeParameterExpression):
             raise ValueError("FreeParameter names must be non empty")
         if not isinstance(name, str):
             raise TypeError("FreeParameter names must be strings")
-        if not name[0].isalpha() and not name[0] == "_":
+        if not name[0].isalpha() and name[0] != "_":
             raise ValueError("FreeParameter names must start with a letter or an underscore")
+        if name in PREDEFINED_VARIABLE_NAMES:
+            raise ValueError(
+                f"FreeParameter names must not be one of the Braket reserved variable names: "
+                f"{PREDEFINED_VARIABLE_NAMES}."
+            )
+        if name in QASM_RESERVED_WORDS:
+            raise ValueError(
+                f"FreeParameter names must not be one of the OpenQASM or OpenPulse keywords: "
+                f"{QASM_RESERVED_WORDS}."
+            )
         self._name = Symbol(name)
 
     def to_dict(self) -> dict: