+++ test/unit_tests/braket/ahs/test_analog_hamiltonian_simulation.py
@@ -70,6 +70,140 @@ def local_detuning():
     )
 
 
+@pytest.fixture
+def ir():
+    return Program.parse_raw_schema(
+        """
+{
+  "braketSchemaHeader": {
+    "name": "braket.ir.ahs.program",
+    "version": "1"
+  },
+  "setup": {
+    "ahs_register": {
+      "sites": [
+        [
+          "0.0",
+          "0.0"
+        ],
+        [
+          "0.0",
+          "0.000003"
+        ],
+        [
+          "0.0",
+          "0.000006"
+        ],
+        [
+          "0.000003",
+          "0.0"
+        ],
+        [
+          "0.000003",
+          "0.000003"
+        ],
+        [
+          "0.000003",
+          "0.000003"
+        ],
+        [
+          "0.000003",
+          "0.000006"
+        ]
+      ],
+      "filling": [
+        1,
+        1,
+        1,
+        1,
+        1,
+        0,
+        0
+      ]
+    }
+  },
+  "hamiltonian": {
+    "drivingFields": [
+      {
+        "amplitude": {
+          "time_series": {
+            "values": [
+              "0.0",
+              "25132700.0",
+              "25132700.0",
+              "0.0"
+            ],
+            "times": [
+              "0.0",
+              "3E-7",
+              "0.0000027",
+              "0.000003"
+            ]
+          },
+          "pattern": "uniform"
+        },
+        "phase": {
+          "time_series": {
+            "values": [
+              "0",
+              "0"
+            ],
+            "times": [
+              "0.0",
+              "0.000003"
+            ]
+          },
+          "pattern": "uniform"
+        },
+        "detuning": {
+          "time_series": {
+            "values": [
+              "-125664000.0",
+              "-125664000.0",
+              "125664000.0",
+              "125664000.0"
+            ],
+            "times": [
+              "0.0",
+              "3E-7",
+              "0.0000027",
+              "0.000003"
+            ]
+          },
+          "pattern": "uniform"
+        }
+      }
+    ],
+    "localDetuning": [
+      {
+        "magnitude": {
+          "time_series": {
+            "values": [
+              "-125664000.0",
+              "125664000.0"
+            ],
+            "times": [
+              "0.0",
+              "0.000003"
+            ]
+          },
+          "pattern": [
+            "0.5",
+            "1.0",
+            "0.5",
+            "0.5",
+            "0.5",
+            "0.5"
+          ]
+        }
+      }
+    ]
+  }
+}
+"""
+    )
+
+
 def test_create():
     mock0 = Mock()
     mock1 = Mock()
@@ -92,6 +226,38 @@ def test_to_ir_empty():
     ahs = AnalogHamiltonianSimulation(register=AtomArrangement(), hamiltonian=hamiltonian)
     problem = ahs.to_ir()
     assert Program.parse_raw(problem.json()) == problem
+    assert problem == Program.parse_raw_schema(problem.json())
+
+
+def test_from_ir(ir):
+    problem = AnalogHamiltonianSimulation.from_ir(ir).to_ir()
+    assert problem == ir
+    assert problem == Program.parse_raw_schema(problem.json())
+
+
+def test_from_ir_empty():
+    empty_ir = Program.parse_raw_schema(
+        """
+{
+  "braketSchemaHeader": {
+    "name": "braket.ir.ahs.program",
+    "version": "1"
+  },
+  "setup": {
+    "ahs_register": {
+      "sites": [],
+      "filling": []
+    }
+  },
+  "hamiltonian": {
+    "drivingFields": [],
+    "localDetuning": []
+  }
+}
+"""
+    )
+    problem = AnalogHamiltonianSimulation.from_ir(empty_ir).to_ir()
+    assert problem == empty_ir
     assert problem == Program.parse_raw_schema(problem.json())
 
 
@@ -141,8 +307,6 @@ def test_discretize(register, driving_field, local_detuning):
     device.properties.paradigm.rydberg.rydbergGlobal.phaseResolution = Decimal("5E-7")
 
     device.properties.paradigm.rydberg.rydbergLocal.timeResolution = Decimal("1E-9")
-    device.properties.paradigm.rydberg.rydbergLocal.commonDetuningResolution = Decimal("2000.0")
-    device.properties.paradigm.rydberg.rydbergLocal.localDetuningResolution = Decimal("0.01")
 
     discretized_ahs = ahs.discretize(device)
     discretized_ir = discretized_ahs.to_ir()
@@ -179,10 +343,10 @@ def test_discretize(register, driving_field, local_detuning):
     }
     local_detuning = discretized_json["hamiltonian"]["localDetuning"][0]["magnitude"]
     assert local_detuning == {
-        "pattern": ["0.50", "1.00", "0.50", "0.50", "0.50", "0.50"],
+        "pattern": ["0.5", "1", "0.5", "0.5", "0.5", "0.5"],
         "time_series": {
             "times": ["0E-9", "0.000003000"],
-            "values": ["-125664000.0", "125664000.0"],
+            "values": ["-125664000", "125664000"],
         },
     }